feat: Retry release fetching
People mess up. They sometimes forget to add assets to a release or
the release is just fetched to quickly after creation and does not
have assets configured yet. In any case, to combat this, the updater
will try a number of previous releases before finally giving up.

The idea is that there should be a matching release at some point
and that the auto-updating nature should not be too dependent on the
latest of the greatest.